1. model backbone tidy up 2. pifpaf debuged
