+++ hyperpose/Model/openpose/model/mbv2_sm_openpose.py
@@ -2,10 +2,11 @@ import tensorflow as tf
 import tensorlayer as tl
 from tensorlayer import layers
 from tensorlayer.models import Model
-from tensorlayer.layers import BatchNorm2d, Conv2d, DepthwiseConv2d, LayerList, MaxPool2d ,SeparableConv2d, UpSampling2d
+from tensorlayer.layers import BatchNorm2d, Conv2d, LayerList ,SeparableConv2d
 from ..utils import tf_repeat
 from ..define import CocoPart,CocoLimb
 from ...common import regulize_loss
+from ...backbones import MobilenetSmall_backbone
 
 class Mobilenetv2_small_Openpose(Model):
     def __init__(self,parts=CocoPart,limbs=CocoLimb,colors=None,n_pos=19,n_limbs=19,num_channels=128,\
@@ -29,7 +30,7 @@ class Mobilenetv2_small_Openpose(Model):
         else:
             self.concat_dim=-1
         if(backbone==None):
-            self.backbone=self.Mobilenetv2_variant(data_format=self.data_format)
+            self.backbone=MobilenetSmall_backbone(data_format=self.data_format)
         else:
             self.backbone=backbone(scale_size=8,pretraining=pretraining,data_format=self.data_format)
         self.init_stage=self.Init_stage(n_confmaps=self.n_confmaps,n_pafmaps=self.n_pafmaps,in_channels=self.backbone.out_channels,data_format=self.data_format)
@@ -99,43 +100,6 @@ class Mobilenetv2_small_Openpose(Model):
 
         return total_loss
 
-    class Mobilenetv2_variant(Model):
-        def __init__(self,data_format="channels_first"):
-            super().__init__()
-            self.data_format=data_format
-            if(self.data_format=="channels_first"):
-                self.concat_dim=1
-            else:
-                self.concat_dim=-1
-            self.out_channels=704
-            self.scale_size=8
-            self.convblock_0=conv_block(n_filter=32,in_channels=3,filter_size=(3,3),strides=(2,2),act=tf.nn.relu,data_format=self.data_format)
-            self.convblock_1=separable_block(n_filter=64,in_channels=32,filter_size=(3,3),strides=(1,1),act=tf.nn.relu,data_format=self.data_format)
-            self.convblock_2=separable_block(n_filter=128,in_channels=64,filter_size=(3,3),strides=(2,2),act=tf.nn.relu,data_format=self.data_format)
-            self.convblock_3=separable_block(n_filter=128,in_channels=128,filter_size=(3,3),strides=(1,1),act=tf.nn.relu,data_format=self.data_format)
-            self.convblock_4=separable_block(n_filter=256,in_channels=128,filter_size=(3,3),strides=(2,2),act=tf.nn.relu,data_format=self.data_format)
-            self.convblock_5=separable_block(n_filter=256,in_channels=256,filter_size=(3,3),strides=(1,1),act=tf.nn.relu,data_format=self.data_format)
-            self.convblock_6=separable_block(n_filter=512,in_channels=256,filter_size=(3,3),strides=(1,1),act=tf.nn.relu,data_format=self.data_format)
-            self.convblock_7=separable_block(n_filter=512,in_channels=512,filter_size=(3,3),strides=(1,1),act=tf.nn.relu,data_format=self.data_format)
-            self.maxpool=MaxPool2d(filter_size=(2,2),strides=(2,2),padding="SAME",data_format=self.data_format)
-            self.upsample=UpSampling2d(scale=2,data_format=self.data_format)
-
-        def forward(self,x):
-            concat_list=[]
-            x=self.convblock_0.forward(x)
-            x=self.convblock_1.forward(x)
-            concat_list.append(self.maxpool.forward(x))
-            x=self.convblock_2.forward(x)
-            x=self.convblock_3.forward(x)
-            concat_list.append(x)
-            x=self.convblock_4.forward(x)
-            x=self.convblock_5.forward(x)
-            x=self.convblock_6.forward(x)
-            x=self.convblock_7.forward(x)
-            concat_list.append(self.upsample.forward(x))
-            x=tf.concat(concat_list,self.concat_dim)
-            return x
-        
     class Init_stage(Model):
         def __init__(self,n_confmaps=19,n_pafmaps=38,in_channels=704,data_format="channels_first"):
             self.n_confmaps=n_confmaps
